+++ internal/docker/docker.go
@@ -2,7 +2,6 @@ package docker
 
 import (
 	"context"
-	"errors"
 	"strings"
 
 	"github.com/docker/docker/api/types"
@@ -11,8 +10,6 @@ import (
 )
 
 func NewClient() (*Client, error) {
-	// TODO: add non-lazy initialisation to ensure that
-	//  the process runs as "inside" container and shares the network.
 	c, err := client.NewClientWithOpts()
 	if err != nil {
 		return nil, err
@@ -24,60 +21,42 @@ type Client struct {
 	*client.Client
 }
 
-func (c *Client) Find(ctx context.Context, instanceNameFilter string) (map[string]struct{}, error) {
+func (c *Client) Scan(ctx context.Context, serviceLabelFilter string) (map[string]string, error) {
 	const statusOK = "running"
 	containers, err := c.ContainerList(ctx, types.ContainerListOptions{
 		Filters: filters.NewArgs(filters.KeyValuePair{
 			Key:   "name",
-			Value: instanceNameFilter,
+			Value: serviceLabelFilter,
 		}),
 	})
 	if err != nil {
 		return nil, err
 	}
 
-	var o = make(map[string]struct{}, len(containers))
+	var o = make(map[string]string)
 	for _, container := range containers {
 		if container.State == statusOK {
-			o[container.ID] = struct{}{}
+			if container.NetworkSettings != nil {
+				for _, network := range container.NetworkSettings.Networks {
+					o[container.ID] = network.IPAddress
+					// use the first found IP for now
+					break
+				}
+			}
 		}
 	}
 	return o, err
 }
 
-func (c *Client) Read(ctx context.Context, id string) (
-	ipAddress string, accessKeyID string, secretAccessKey string,
-	err error,
-) {
-	info, err := c.ContainerInspect(ctx, id)
+func (c *Client) Read(ctx context.Context, instanceID string) (string, string, error) {
+	info, err := c.ContainerInspect(ctx, instanceID)
 	if err != nil {
-		return "", "", "", err
+		return "", "", err
 	}
 
-	settings := info.NetworkSettings
-	if settings == nil {
-		return "", "", "",
-			errors.New("cannot find network configuration for the instance " + id)
-	}
+	accessKeyID, secretAccessKey := readAccessCredentialsFromEnv(info.Config.Env)
 
-	for _, network := range settings.Networks {
-		ipAddress = network.IPAddress
-
-		// select the first ip
-		// TODO: identify the Gateway process and use it to select common network
-		if ipAddress != "" {
-			break
-		}
-	}
-
-	// fallback: if the docker process runs in the default network
-	if ipAddress == "" {
-		ipAddress = settings.DefaultNetworkSettings.IPAddress
-	}
-
-	accessKeyID, secretAccessKey = readAccessCredentialsFromEnv(info.Config.Env)
-
-	return ipAddress, accessKeyID, secretAccessKey, nil
+	return accessKeyID, secretAccessKey, nil
 }
 
 func readAccessCredentialsFromEnv(envVars []string) (accessKeyID, secretAccessKey string) {